tests/wpt/web-platform-tests/forced-colors-mode/forced-colors-mode-11.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>
  Forced colors mode - forced-colors media query.
  Tests that styles defined inside forced-color media query
  are preserved in forced colors mode if forced-color-adjust
  is none.
</title>
<link rel="help" href="https://www.w3.org/TR/css-color-adjust-1/#forced-colors-properties">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  @media (forced-colors: none) {
    p {
      background-color: rgb(255, 0, 0);
      color: rgb(255, 255, 255);
    }
  }
  @media (forced-colors: active) {
    p {
      background-color: rgb(0, 0, 255);
      color: rgb(0, 128, 0);
      forced-color-adjust: none;
    }
  }
</style>
</head>
<body>
  <p id="p">
    This text should be green in forced colors mode.
  </p>
</body>

<script>
  test(function() {
    assert_equals(getComputedStyle(p).backgroundColor, "rgb(0, 0, 255)");
    assert_equals(getComputedStyle(p).color, "rgb(0, 128, 0)");
  }, "Checks that styles defined in forced-colors are preserved in forced colors mode.");
</script>
